refactoring: context instead of props drilling
