+++ src/Components/SavedSearchAlert/Utils/savedSearchCriteria.tsx
@@ -8,8 +8,6 @@ import {
   SavedSearchEntity,
   SearchCriteriaAttributes,
 } from "Components/SavedSearchAlert/types"
-import { paramsToSnakeCase } from "Components/ArtworkFilter/Utils/urlBuilder"
-import qs from "qs"
 
 export const isDefaultValue = (
   paramName: string,
@@ -68,20 +66,4 @@ export const getSearchCriteriaFromFilters = (
     ...allowedFilters,
     ...defaultCriteria,
   }
-}
-
-export const searchCriteriaHref = (
-  artistSlug: string,
-  criteria: SearchCriteriaAttributes
-) => {
-  const allowedCriteriaValues = Object.fromEntries(
-    Object.entries(criteria).filter(([key, _]) => {
-      // artistSlug is passed separately
-      if (key === "artistIDs") return false
-      return allowedSearchCriteriaKeys.includes(key)
-    })
-  )
-  const queryParams = qs.stringify(paramsToSnakeCase(allowedCriteriaValues))
-
-  return `/artist/${artistSlug}?${queryParams}`
 }
